Title: 트리에서 공 떨어뜨리기, Time: 115ms, Memory: 1MB, Status: Wrong Answer - CodeTree

+++ 240309/트리에서 공 떨어뜨리기/drop-a-ball-in-a-tree.cpp
@@ -55,11 +55,11 @@ int find_next(int number){
         //현재 k가 홀수인 경우
         //현재 갈림길에서 왼쪽으로 이동.
         next_root = find_sub_tree(left);
-        curr_k = curr_k / 2 + 1;
+        curr_k = (long long)(curr_k / 2 + 1);
     }
     else{
         next_root = find_sub_tree(right);
-        curr_k = curr_k / 2;
+        curr_k = (long long)(curr_k / 2);
     }
     return next_root;
